Changed the default text from "Web Page" to "Web Resource"
I found that Excel files default to "Web Page" because the file type is not in the RESOURCE_MAPPING section. I did a pull request to update the RESOURCE_MAPPING section of https://github.com/GSA/ckanext-geodatagov/blob/master/ckanext/geodatagov/plugins.py

I think "Web Resource" is more generic and less confusing if the file format is not found in the RESOURCE_MAPPING section.