Temporarily copy nose tests to subdir
Preserve the existing nose tests temporarily. This allows us to keep the
existing tests running and checking extension code, while allowing us to
modernize tests for pytest.

pytest is supported for CKAN 2.8 and CKAN 2.9, so once we are confident 2.8
pytests are running correctly, the nose tests can be dropped entirely.