Fix facet filter tests
Not sure if this is really correct. With our facet_list override, we were seeing
no facets filters rendered. The difference was between the nose tests and the
CKAN 2.8 tests, which doesn't really make sense, since both are using the same
templates.

Not seeing a clear reason why we're overriding the facet_list template to begin
with, let's move back to the upstream template, which at least is recovering
technical debt.

Also update the test code to be explicit about what we're testing (appearance of
specific facet filters). The comment referred to `groups`, but this facet filter
does not appear, and is not being asserted. Removed.