Fix get_bureau_info tests
Convert helper tests to functions instead of classes. pytest prefers function
based tests and then using setup/teardown functions as fixtures. pytest _is_
compatible with `unittest`, but we're not using unittest. It seems that when
using vanilla classes with `__init__`, suddenly the class isn't being considered
a test suite and the tests are not running.

I rewrote the test class as functions with an assert helper which feels a bit
more natural.