Remove HTML metadata link
This is super legacy code, and something that data.gov should not be responsible for. While we accept ISO and CSDGM metadata standards, we are not experts in them and should not be trying to display them in custom HTML formats. Hopefully the recent work to link to  link to Geoplatform (https://github.com/GSA/data.gov/issues/3478) will point to the experts in geospatial metadata.